Add index page for blog list
